+++ cli_auth_no_sigtstp.go
@@ -1,0 +1,11 @@
+//go:build !unix
+
+package jwtauth
+
+import (
+	"os"
+)
+
+// authHalts are the signals we want to interrupt our auth callback on.
+// SIGTSTP is omitted for Windows.
+var authHalts = []os.Signal{os.Interrupt, os.Kill}